fix SIKE ct function
